+++ awx/api/templates/api/bulk_view.md
@@ -1,0 +1,3 @@
+# Bulk Actions
+
+This endpoint lists available bulk action APIs.
